Removed dead code
