+++ src/com/morgan/design/utils/Utils.java
@@ -1,13 +1,9 @@
 package com.morgan.design.utils;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
-import android.os.Build;
 import android.os.Vibrator;
 import android.text.Spanned;
 import android.view.View;
@@ -22,8 +18,6 @@ import com.morgan.design.analytics.GoogleAnalyticsActivity;
 import com.morgan.design.helpers.Constants;
 
 public class Utils {
-
-	private final static Logger LOG = LoggerFactory.getLogger(Utils.class);
 
 	public static void createEmailIntenet(final Activity activity, final String title, final String type, final String subject,
 			final String[] toAddresses, final Spanned body) {
@@ -138,14 +132,6 @@ public class Utils {
 	public static void openImprovement(final GoogleAnalyticsActivity activity) {
 		Utils.sendEmail(activity.getActivity(), "Send mail...", "Who's Making The Brew : Improvement", Constants.MY_EMAIL);
 		activity.trackPageView("/OpenImprovmentEmail|" + activity.getClass().getCanonicalName());
-	}
-
-	public static void logBuildDetails() {
-		LOG.debug("##########################################");
-		LOG.debug(Build.VERSION.CODENAME);
-		LOG.debug(Build.VERSION.RELEASE);
-		LOG.debug(Build.VERSION.SDK);
-		LOG.debug("##########################################");
 	}
 
 }
